+++ app/test/e2e/collection/collection-get.spec.js
@@ -1,5 +1,6 @@
 const nock = require('nock');
 const chai = require('chai');
+const config = require('config');
 const Collection = require('models/collection.model');
 const { USERS } = require('../utils/test.constants');
 const { createCollection, mockGetUserFromToken } = require('../utils/helpers');
@@ -22,6 +23,43 @@ describe('Get collections', () => {
         requester = await getTestServer();
 
         await Collection.deleteMany({}).exec();
+    });
+
+    describe('Test pagination links', () => {
+        it('Get collections without referer header should be successful and use the request host', async () => {
+            mockGetUserFromToken(USERS.USER);
+            const response = await requester
+                .get(`/api/v1/collection`)
+                .set('Authorization', `Bearer abcd`)
+                .send();
+
+            response.status.should.equal(200);
+            response.body.should.have.property('data').and.be.an('array');
+            response.body.should.have.property('links').and.be.an('object');
+            response.body.links.should.have.property('self').and.equal(`http://127.0.0.1:${config.get('service.port')}/v1/collection?page[number]=1&page[size]=9999999`);
+            response.body.links.should.have.property('prev').and.equal(`http://127.0.0.1:${config.get('service.port')}/v1/collection?page[number]=1&page[size]=9999999`);
+            response.body.links.should.have.property('next').and.equal(`http://127.0.0.1:${config.get('service.port')}/v1/collection?page[number]=1&page[size]=9999999`);
+            response.body.links.should.have.property('first').and.equal(`http://127.0.0.1:${config.get('service.port')}/v1/collection?page[number]=1&page[size]=9999999`);
+            response.body.links.should.have.property('last').and.equal(`http://127.0.0.1:${config.get('service.port')}/v1/collection?page[number]=1&page[size]=9999999`);
+        });
+
+        it('Get collections with referer header should be successful and use that header on the links on the response', async () => {
+            mockGetUserFromToken(USERS.USER);
+            const response = await requester
+                .get(`/api/v1/collection`)
+                .set('Authorization', `Bearer abcd`)
+                .set('referer', `https://potato.com/get-me-all-the-data`)
+                .send();
+
+            response.status.should.equal(200);
+            response.body.should.have.property('data').and.be.an('array');
+            response.body.should.have.property('links').and.be.an('object');
+            response.body.links.should.have.property('self').and.equal('http://potato.com/v1/collection?page[number]=1&page[size]=9999999');
+            response.body.links.should.have.property('prev').and.equal('http://potato.com/v1/collection?page[number]=1&page[size]=9999999');
+            response.body.links.should.have.property('next').and.equal('http://potato.com/v1/collection?page[number]=1&page[size]=9999999');
+            response.body.links.should.have.property('first').and.equal('http://potato.com/v1/collection?page[number]=1&page[size]=9999999');
+            response.body.links.should.have.property('last').and.equal('http://potato.com/v1/collection?page[number]=1&page[size]=9999999');
+        });
     });
 
     it('Get collections without being authenticated should return a 401 `Unauthorized`', async () => {